+++ src/test/java/com/braintreegateway/unittest/PaymentMethodNonceTest.java
@@ -187,6 +187,13 @@ public class PaymentMethodNonceTest {
                 "  <details>" +
                 "    <email>jane.doe@paypal.com</email>" +
                 "    <correlation-id>46676383-b632-4b80-8cfd-d7a35d960888</correlation-id>" +
+                "    <payer-info>" +
+                "      <email>jane2.doe@paypal.com</email>" +
+                "      <first-name>first</first-name>" +
+                "      <last-name>last</last-name>" +
+                "      <payer-id>pay-123</payer-id>" +
+                "      <country-code>US</country-code>" +
+                "    </payer-info>" +
                 "  </details>" +
                 "</payment-method-nonce>";
 
@@ -199,5 +206,11 @@ public class PaymentMethodNonceTest {
         assertEquals(false, paymentMethodNonce.isConsumed());
         assertNotNull(paymentMethodNonce.getDetails());
         assertEquals("jane.doe@paypal.com", paymentMethodNonce.getDetails().getEmail());
+
+        assertEquals("jane2.doe@paypal.com", paymentMethodNonce.getDetails().getPayerInfo().getEmail());
+        assertEquals("first", paymentMethodNonce.getDetails().getPayerInfo().getFirstName());
+        assertEquals("last", paymentMethodNonce.getDetails().getPayerInfo().getLastName());
+        assertEquals("pay-123", paymentMethodNonce.getDetails().getPayerInfo().getPayerId());
+        assertEquals("US", paymentMethodNonce.getDetails().getPayerInfo().getCountryCode());
     }
 }